# myblog/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gin+gorm+jwt+cors+viper等等",
        "title": "myblog gin+gorm",
        "contact": {
            "name": "go-swagger帮助文档",
            "url": "https://github.com/swaggo/swag/blob/master/README_zh-CN.md"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/board/add": {
            "post": {
                "description": "登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "parameters": [
                    {
                        "description": "登录",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
...
        "/api/lifecomment/add": {
            "post": {
                "description": "前台添加一个生活评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LifeComment"
                ],
                "summary": "前台添加生活评论",
                "parameters": [
                    {
                        "description": "添加生活评论",
                        "name": "lifecomment",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.LifeComment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/lifecomment/delete/{id}": {
            "delete": {
                "description": "后台删除一个动态评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Life"
                ],
                "summary": "后台删除动态评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要删除的动态评论id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/lifecomment/list": {
            "get": {
                "description": "后台查看动态评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LifeComment"
                ],
                "summary": "后台查看动态评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页尺寸",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pagenum",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
        "/api/lifecomment/{id}": {
            "get": {
                "description": "前台查看动态评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LifeComment"
                ],
                "summary": "前台查看动态评论",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分页尺寸",
                        "name": "pagesize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "pagenum",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "life的id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Reponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Life": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "img": {
                    "type": "string"
                },
                "read_count": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "model.LifeComment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "lid": {
                    "type": "integer"
                },
                "life": {
                    "description": "关联Life，并将Lid重写为外键",
                    "$ref": "#/definitions/model.Life"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "model.Reponse": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}